+++ src/utils/command_utils/exit_utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   exit_utils.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tboldrin <tboldrin@student.42.fr>          +#+  +:+       +#+        */
+/*   By: rciaze <rciaze@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/04 15:41:50 by tboldrin          #+#    #+#             */
-/*   Updated: 2023/10/04 19:01:51 by tboldrin         ###   ########.fr       */
+/*   Updated: 2023/10/05 18:10:39 by rciaze           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -62,7 +62,7 @@ void	ft_exit(int code, bool msg)
 	exit(code);
 }
 
-void	update_err_code_exit(char *origin_code, int code_err)
+void	update_err_code_exit(char *origin_code, long long int code_err)
 {
 	char	*str;
